Fix PHP Notice on Empty Fields
When a Cvent record (e.g. a Registration) has empty data (e.g. the 
Middle Name), the field is not being returned in the API result. We 
should check for this emptiness before setting up results to avoid a PHP 
Notice on the undefined array key.

A caveat of this is that we cannot detect an empty field vs. a field 
that is not actually defined on the object, but I think it's the best we 
can do considering this is a limitation of the API, not the wrapper.