Scrub Creds from Auth Errors
Restructure error-handling a bit to ensure we can throw/catch auth 
errors and scrub the account info. out of the message. Also added 
`$throw_fault` bool to the `_call` method to trigger SoapFault sequence 
even when the request was fine. Not used or implemented in production, 
leaving it there for testing.